image-acquirer/app/requirements.txt
opencv-python
Flask
ultralytics
torchvision
lapx>=0.5.2
numpy==1.23
paho-mqtt
dill
tensorrt
